+++ getdir.php
@@ -12,26 +12,20 @@ if ($action == 'i' || $action == 's' || $action == 'o') {
         if ($fsockopen != false) {
             if (file_exists($repo)) {
                 exec('chmod -R 777 .'); exec('rm -rf '.$repo);
-            }
-            if ($branch != '') {
+            } if ($branch != '') {
                 exec('git clone -b '.$branch.' '.$host.'/'.$user.'/'.$repo);
             } else {
                 exec('git clone '.$host.'/'.$user.'/'.$repo);
-            }
-            chmod($repo, 0777); exec('chmod -R 777 .');
+            } chmod($repo, 0777); exec('chmod -R 777 .');
         }
     }
 } elseif ($action == 'r' || $action == 'p' || $action == 'm') {
     if ($pkg != "" && $repo != "" && $user != "") {
-        if (file_exists($pkg)) {
-            exec('chmod -R 777 .'); exec('rm -rf '.$pkg);
-        }
+        if (file_exists($pkg)) { exec('chmod -R 777 .'); exec('rm -rf '.$pkg); }
         header('Location: getdir.php?key=i&host='.$host.'&pkg=from&repo='.$repo.'&branch='.$branch.'&user='.$user);
     }
 } elseif ($action == 'd' || $action == 'u' || $action == 'x') {
     if ($pkg != "" && $repo == 'from' && $user == 'here') {
-        if (file_exists($pkg)) {
-            exec('chmod -R 777 .'); exec('rm -rf '.$pkg);
-        }
+        if (file_exists($pkg)) { exec('chmod -R 777 .'); exec('rm -rf '.$pkg); }
     }
 }
